+++ lib/features/auth/domain/log_in_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:noti_flutter/features/auth/data/repositories/auth_repository.dart';
-import 'package:noti_flutter/features/fire_store/auth_repository.dart';
+import 'package:noti_flutter/features/fire_store/fire_store_repository.dart';
 import 'package:noti_flutter/model/user_model.dart';
 
 final logInUsecaseProvider = Provider((ref) {
