+++ lib/features/auth/domain/sign_up_service.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:noti_flutter/features/auth/data/dto/sign_up_dto.dart';
 import 'package:noti_flutter/features/auth/data/repositories/auth_repository.dart';
-import 'package:noti_flutter/features/fire_store/auth_repository.dart';
+import 'package:noti_flutter/features/fire_store/fire_store_repository.dart';
 import 'package:noti_flutter/model/user_model.dart';
 
 final signUpServiceProvider = Provider<SignUpService>((ref) {
